Simplify newline in log
On my system echo does not interpret the \n char and it is printed to the file as is. This leads to other errors when converting later. Usually -e would be used, but that is also not available on every system from what I read.
echo without any param does the trick for me though.